Fixed the Twitpocalypse by converting all uint/UInt to qlonglong/QLongLong http://news.cnet.com/8301-17939_109-10264002-2.html
svn path=/trunk/extragear/network/choqok/; revision=981277

+++ src/account.cpp
@@ -48,12 +48,12 @@ Account::Account( const Account & account )
     setServiceType( account.serviceType(), account.homepage() );
 }
 
-uint Account::userId() const
+qulonglong Account::userId() const
 {
     return mUserId;
 }
 
-void Account::setUserId( uint id )
+void Account::setUserId( qulonglong id )
 {
     mUserId = id;
 }
