+++ src/search.h
@@ -50,14 +50,14 @@ public:
     QMap<int, QPair<QString, bool> > getSearchTypes();
 
 private:
-    virtual KUrl buildUrl( QString query, int option, uint sinceStatusId = 0, uint count = 0, uint page = 1 );
+    virtual KUrl buildUrl( QString query, int option, qulonglong sinceStatusId = 0, qulonglong count = 0, qulonglong page = 1 );
 
 public slots:
     virtual void requestSearchResults( QString query,
                                        int option,
-                                       uint sinceStatusId = 0,
-                                       uint count = 0,
-                                       uint page = 1 );
+                                       qulonglong sinceStatusId = 0,
+                                       qulonglong count = 0,
+                                       qulonglong page = 1 );
 
 protected slots:
     virtual void searchResultsReturned( KJob *job );
@@ -71,7 +71,7 @@ protected:
     // The QString in the QPair is a human readable string describing what the type searches for. The boolean value
     // determines whether or not the search type is traversable (if the forward and back buttons should be displayed).
     QMap<int, QPair<QString, bool> > mSearchTypes;
-    uint mSinceStatusId;
+    qulonglong mSinceStatusId;
     QString mSearchUrl;
     Account* mAccount;
 };
